feat(referral): add Detail section header to hardware records list
Add a "Details" heading above the hardware records list using
$headingLg typography. Also reorganize HardwareSalesRewardHeader
into its own directory for better code organization.
