+++ packages/kit/src/views/ReferFriends/pages/HardwareSalesReward/components/HardwareSalesRewardHeader/index.tsx
@@ -1,0 +1,3 @@
+export { HardwareSalesRewardHeader } from './HardwareSalesRewardHeader';
+export { StatCard } from './StatCard';
+export type { IStatCardProps } from './StatCard';
